feat: Add Tech Stack component and embed card for certifications
- Implemented Tech Stack component showcasing various technologies used in the project.
- Created EmbedCard component to dynamically load and display badges from Credly.
- Developed navigation links for the website with active link highlighting.
- Added responsive navigation bar components for different screen sizes.
- Integrated social media links with tooltips for better user experience.
- Introduced theme toggler for light and dark mode switching.
- Added certification and expertise data for better content representation.
- Configured ESLint for code quality and consistency.
