+++ actions/wakatime.ts
@@ -2,13 +2,11 @@
 import axios from "axios";
 
 export const FetchWakaTimeStats = async () => {
-  try {
-    const response = await axios.get(
-      "https://wakatime.com/api/v1/users/froilanimnida/stats?including_today=true",
-    );
-    return response.data;
-  } catch (error) {
-    console.error(error);
-    return null;
-  }
+    try {
+        const response = await axios.get("https://wakatime.com/api/v1/users/froilanimnida/stats?including_today=true");
+        return response.data;
+    } catch (error) {
+        console.error(error);
+        return null;
+    }
 };
