+++ app/experience/page.tsx
@@ -1,11 +1,7 @@
-import React from "react";
-
 export default function Experience() {
-  return (
-    <div className="w-11/12 max-w-7xl flex flex-col gap-2 justify-start items-start">
-      <h1 className="font-bold text-3xl lg:font-extrabold md:text-4xl lg:text-5xl xl:text-8xl">
-        Experience
-      </h1>
-    </div>
-  );
+    return (
+        <div className="w-11/12 max-w-7xl flex flex-col gap-2 justify-start items-start">
+            <h1 className="font-bold text-3xl lg:font-extrabold md:text-4xl lg:text-5xl xl:text-8xl">Experience</h1>
+        </div>
+    );
 }
