+++ components/navigation-bar/navbar.tsx
@@ -1,13 +1,13 @@
 import Logo from "@/components/logo";
-import Links from "@/components/NavigationBar/Links";
-import Socials from "@/components/NavigationBar/Socials";
+import NavLinks from "@/components/navigation-bar/nav-links";
+import Socials from "@/components/navigation-bar/socials";
 
 const NavBar = () => {
     return (
         <>
             <div className="hidden md:flex w-full md:w-[95%] lg:w-[85%] xl:w-[80%] items-center flex-row justify-between">
                 <Logo />
-                <Links />
+                <NavLinks />
                 <Socials />
             </div>
         </>
