+++ components/navigation-bar/socials.tsx
@@ -1,5 +1,5 @@
 import Link from "next/link";
-import { ThemeToggler } from "@/components/ThemeToggler";
+import { ThemeToggler } from "@/components/theme-toggler";
 import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from "@/components/ui";
 
 export const metadata = {
