+++ data/certification-list-preview.ts
@@ -1,0 +1,30 @@
+export const CertificationsListsPreview = [
+    {
+        id: 0,
+        title: "Python Essentials 1",
+        description: "Cisco Networking Academy",
+        link: "https://www.credly.com/badges/b8e988b8-ec7d-4209-bf2e-2849a92ef464/public_url",
+        badge_id: "b8e988b8-ec7d-4209-bf2e-2849a92ef464",
+    },
+    {
+        id: 1,
+        title: "Python Essentials 2",
+        description: "Cisco Networking Academy",
+        link: "https://www.credly.com/badges/2bd4e85e-4cf0-460d-a2b6-2e8facfc6d6b/public_url",
+        badge_id: "2bd4e85e-4cf0-460d-a2b6-2e8facfc6d6b",
+    },
+    {
+        id: 2,
+        title: "Operating Systems Basics",
+        description: "Cisco Networking Academy",
+        link: "https://www.credly.com/badges/6bcbf2ef-2079-41f6-a4d7-9305871b67d2/public_url",
+        badge_id: "6bcbf2ef-2079-41f6-a4d7-9305871b67d2",
+    },
+    {
+        id: 3,
+        title: "JavaScript Essentials 1",
+        description: "Cisco Networking Academy",
+        link: "https://www.credly.com/badges/88406fa2-a934-41c2-8fbd-5de9fefe2368/public_url",
+        badge_id: "88406fa2-a934-41c2-8fbd-5de9fefe2368",
+    },
+];
