+++ data/expertise-list.ts
@@ -1,0 +1,88 @@
+export const ExpertiseList = [
+    {
+        name: "Got Python Essentials Badge from Cisco Networking Academy",
+        dateObtained: "JUNE 22, 2024",
+        content:
+            "In light with my passion for programming, I took the Python Programming Essentials course from Cisco Networking Academy and earned a badge for it. I learned the basics of Python programming, including data types, control structures, and functions. I also learned how to use Python to create scripts that automate tasks and solve real-world problems. I am confident that I can use my Python programming skills to contribute to any project that requires automation or data analysis. Additionally, I learned various skills such as analytical thinking, algorithmic thinking, and problem-solving skills that will be useful in my future endeavors.",
+        badge: [
+            "Python",
+            "Cisco Networking Academy",
+            "Basic Python Programming",
+            "Entry-level",
+            "Analytical Thinking",
+            "Algorithmic Thinking",
+        ],
+    },
+    {
+        name: "Got the 2nd Badge of Python Programming from Cisco Networking Academy",
+        dateObtained: "JULY 10, 2024",
+        content:
+            "Prior finishing the Python Programming Essentials 1 course. I took Python Programming Essentials 2 course from Cisco Networking Academy and earned a badge for it. Since it gave me a more deeper understanding of Python programming and help reinforce the things that I learned from the first course. I learned more about Python classes, Python exceptions, and Python file handling.",
+        badge: [
+            "Python Classes",
+            "Python Exceptions",
+            "Python Functions",
+            "Python Inheritance",
+            "Python Libraries",
+            "Python Modules",
+            "Python PIP",
+            "Python Packages",
+            "Python Generators",
+            "Python Objects",
+            "Python Polymorphism",
+        ],
+    },
+    {
+        name: "Operating Systems: The Basics from Cisco Networking Academy",
+        dateObtained: "JULY 17, 2024",
+        content:
+            "I took the Operating Systems: The Basics course from Cisco Networking Academy and earned a badge for it. I learned the basics of operating systems, including the functions of an operating system, the types of operating systems, and the components of an operating system. I also learned how to use an operating system to manage files, run applications, and connect to networks. I am confident that I can use my knowledge of operating systems to troubleshoot problems, optimize performance, and secure systems.",
+        badge: ["Android", "Cisco Networking Academy", "Linux", "iOS", "Windows"],
+    },
+    {
+        name: "JavaScript Essentials 1 from Cisco Networking Academy",
+        dateObtained: "JULY 31, 2024",
+        content:
+            "I took the JavaScript Essentials 1 course from Cisco Networking Academy and earned a badge for it. I learned the basics of JavaScript programming, including data types, control structures, and functions. I also learned how to use JavaScript to create interactive web pages and web applications. I am confident that I can use my JavaScript programming skills to create dynamic and responsive user interfaces.",
+        badge: [
+            "JavaScript",
+            "Cisco Networking Academy",
+            "Basic JavaScript Programming",
+            "Entry-level",
+            "Interactive Web Pages",
+            "Web Applications",
+            "Recursion",
+            "Loops",
+            "Functions",
+            "Data Types",
+            "Control Structures",
+            "Comments",
+            "Operators",
+            "Type Casting",
+            "Errors",
+            "User Interaction",
+        ],
+    },
+    {
+        name: "JavaScript Essentials 2 from Cisco Networking Academy",
+        dateObtained: "SEPTEMBER 30, 2024",
+        content:
+            "I took the JavaScript Essentials 2 course from Cisco Networking Academy and earned a badge for it. I learned more about JavaScript programming, including arrays, objects, and functions. I also learned how to use JavaScript to create interactive web pages and web applications. I am confident that I can use my JavaScript programming skills to create dynamic and responsive user interfaces.",
+        badge: [
+            "JavaScript",
+            "Cisco Networking Academy",
+            "Intermediate JavaScript Programming",
+            "Arrays Techniques",
+            "Prototype",
+            "Getters",
+            "Enumerating Properties",
+            "Deep Cloning",
+            "Setters",
+            "Testing Techniques",
+            "Property Management",
+            "Inheritance",
+            "Sets",
+            "Maps",
+        ],
+    },
+];
