+++ next.config.js
@@ -1,21 +1,21 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  images: {
-    dangerouslyAllowSVG: true,
-    remotePatterns: [
-      {
-        protocol: "https",
-        hostname: "i.scdn.co",
-        port: "",
-        pathname: "/image/**",
-      },
-      {
-        protocol: "https",
-        hostname: "roadmap.sh",
-        port: "",
-      },
-    ],
-  },
+    images: {
+        dangerouslyAllowSVG: true,
+        remotePatterns: [
+            {
+                protocol: "https",
+                hostname: "i.scdn.co",
+                port: "",
+                pathname: "/image/**",
+            },
+            {
+                protocol: "https",
+                hostname: "roadmap.sh",
+                port: "",
+            },
+        ],
+    },
 };
 
 module.exports = nextConfig;
