+++ package.json
@@ -1,51 +1,54 @@
 {
-  "name": "froilanimnida",
-  "version": "2.0.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "format": "prettier --write .",
-    "format:check": "prettier --check ."
-  },
-  "dependencies": {
-    "@radix-ui/react-dropdown-menu": "^2.1.7",
-    "@radix-ui/react-hover-card": "^1.1.5",
-    "@radix-ui/react-icons": "^1.3.2",
-    "@radix-ui/react-progress": "^1.1.3",
-    "@radix-ui/react-separator": "^1.1.1",
-    "@radix-ui/react-slot": "^1.2.0",
-    "@radix-ui/react-tooltip": "^1.1.8",
-    "@radix-ui/themes": "^3.2.1",
-    "@types/node": "22.14.1",
-    "@types/react": "18.3.12",
-    "@types/react-dom": "18.3.1",
-    "@vercel/analytics": "^1.5.0",
-    "axios": "^1.8.2",
-    "class-variance-authority": "^0.7.1",
-    "clsx": "^2.1.1",
-    "eslint": "9.24.0",
-    "eslint-config-next": "15.2.3",
-    "lucide-react": "^0.501.0",
-    "next": "^15.2.4",
-    "next-themes": "^0.3.0",
-    "react": "18.3.1",
-    "react-dom": "18.3.1",
-    "react-icons": "^5.5.0",
-    "react-markdown": "^10.1.0",
-    "react-syntax-highlighter": "^15.6.1",
-    "sharp": "^0.34.0",
-    "tailwind-merge": "^3.0.2",
-    "tailwindcss-animate": "^1.0.7",
-    "typescript": "5.8.2"
-  },
-  "devDependencies": {
-    "@tailwindcss/postcss": "^4.1.4",
-    "postcss": "^8.5.3",
-    "prettier": "3.5.3",
-    "querystring": "^0.2.1",
-    "tailwindcss": "^4.1.4"
-  }
+    "name": "froilanimnida",
+    "version": "2.0.0",
+    "private": true,
+    "scripts": {
+        "dev": "next dev",
+        "build": "next build",
+        "start": "next start",
+        "lint": "next lint",
+        "format": "prettier --write .",
+        "format:check": "prettier --check ."
+    },
+    "dependencies": {
+        "@radix-ui/react-dropdown-menu": "^2.1.7",
+        "@radix-ui/react-hover-card": "^1.1.5",
+        "@radix-ui/react-icons": "^1.3.2",
+        "@radix-ui/react-progress": "^1.1.3",
+        "@radix-ui/react-separator": "^1.1.1",
+        "@radix-ui/react-slot": "^1.2.0",
+        "@radix-ui/react-tooltip": "^1.1.8",
+        "@radix-ui/themes": "^3.2.1",
+        "@types/node": "22.14.1",
+        "@types/react": "18.3.12",
+        "@types/react-dom": "18.3.1",
+        "@vercel/analytics": "^1.5.0",
+        "axios": "^1.8.2",
+        "class-variance-authority": "^0.7.1",
+        "clsx": "^2.1.1",
+        "eslint-config-next": "15.2.3",
+        "lucide-react": "^0.501.0",
+        "next": "^15.2.4",
+        "next-themes": "^0.3.0",
+        "react": "18.3.1",
+        "react-dom": "18.3.1",
+        "react-icons": "^5.5.0",
+        "react-markdown": "^10.1.0",
+        "react-syntax-highlighter": "^15.6.1",
+        "sharp": "^0.34.0",
+        "tailwind-merge": "^3.0.2",
+        "tailwindcss-animate": "^1.0.7",
+        "typescript": "5.8.2"
+    },
+    "devDependencies": {
+        "@eslint/js": "^9.26.0",
+        "@next/eslint-plugin-next": "^15.3.2",
+        "@tailwindcss/postcss": "^4.1.4",
+        "@types/eslint__eslintrc": "^2.1.2",
+        "eslint": "^8.57.1",
+        "postcss": "^8.5.3",
+        "prettier": "3.5.3",
+        "querystring": "^0.2.1",
+        "tailwindcss": "^4.1.4"
+    }
 }
